Move BookScanner from site_pipeline to io
